style: use serialize
